Rebase the two node interfacers onto a single parent
